+++ src/lib/pages/contract-details/components/InstantiateInfo.tsx
@@ -1,15 +1,27 @@
 import { Box, chakra, Divider, Flex, Text } from "@chakra-ui/react";
 
+import type { ContractData } from "../types";
 import { useGetAddressType } from "lib/app-provider";
 import { Copier } from "lib/components/copy";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { LabelText } from "lib/components/LabelText";
-import type { ContractData, Option } from "lib/types";
+import { Loading } from "lib/components/Loading";
+import type { Option } from "lib/types";
 import { formatUTC, dateFromNow } from "lib/utils";
 import { getAddressTypeText } from "lib/utils/address";
 
 interface InstantiateInfoProps {
-  contractData: ContractData;
+  chainId: ContractData["chainId"];
+  codeLocalInfo: ContractData["codeLocalInfo"];
+  contractDetail: ContractData["contractDetail"];
+  contractCw2Info: ContractData["contractCw2Info"];
+  initTxHash: ContractData["initTxHash"];
+  initProposalId: ContractData["initProposalId"];
+  initProposalTitle: ContractData["initProposalTitle"];
+  createdHeight: ContractData["createdHeight"];
+  createdTime: ContractData["createdTime"];
+  rawContractResponse: ContractData["rawContractResponse"];
+  isLoading: boolean;
 }
 
 const Container = chakra(Flex, {
@@ -95,21 +107,28 @@ const InitRender = ({
 };
 
 export const InstantiateInfo = ({
-  contractData: {
-    contractCw2Info,
-    instantiateInfo,
-    chainId,
-    codeInfo,
-    initTxHash,
-    initProposalId,
-    initProposalTitle,
-    createdHeight,
-    createdTime,
-  },
+  chainId,
+  codeLocalInfo,
+  contractDetail,
+  contractCw2Info,
+  initTxHash,
+  initProposalId,
+  initProposalTitle,
+  createdHeight,
+  createdTime,
+  rawContractResponse,
+  isLoading,
 }: InstantiateInfoProps) => {
   const getAddressType = useGetAddressType();
 
-  if (!instantiateInfo) {
+  if (isLoading)
+    return (
+      <Container>
+        <Loading />
+      </Container>
+    );
+
+  if (!contractDetail) {
     return (
       <Container>
         <Text variant="body2" color="text.dark">
@@ -119,8 +138,8 @@ export const InstantiateInfo = ({
     );
   }
 
-  const instantiatorType = getAddressType(instantiateInfo.instantiator);
-  const adminType = getAddressType(instantiateInfo.admin);
+  const instantiatorType = getAddressType(contractDetail.instantiator);
+  const adminType = getAddressType(contractDetail.admin);
 
   return (
     <Container w={{ base: "full", md: "auto" }}>
@@ -128,10 +147,11 @@ export const InstantiateInfo = ({
         <LabelText flex="1" label="Network">
           {chainId}
         </LabelText>
-        <LabelText flex="1" label="From Code" helperText1={codeInfo?.name}>
+
+        <LabelText flex="1" label="From Code" helperText1={codeLocalInfo?.name}>
           <ExplorerLink
             type="code_id"
-            value={instantiateInfo.codeId}
+            value={contractDetail.codeId.toString()}
             showCopyOnHover
             fixedHeight
           />
@@ -149,7 +169,8 @@ export const InstantiateInfo = ({
             </Text>
           )}
         </LabelText>
-        {instantiateInfo.admin ? (
+
+        {contractDetail.admin ? (
           <LabelText
             flex="1"
             label="Admin Address"
@@ -157,7 +178,7 @@ export const InstantiateInfo = ({
           >
             <ExplorerLink
               type={adminType}
-              value={instantiateInfo.admin}
+              value={contractDetail.admin}
               showCopyOnHover
               fixedHeight
             />
@@ -170,6 +191,7 @@ export const InstantiateInfo = ({
           </LabelText>
         )}
       </Flex>
+
       <Divider border="1px solid" borderColor="gray.700" />
       {createdHeight ? (
         <LabelText
@@ -196,7 +218,7 @@ export const InstantiateInfo = ({
         >
           <ExplorerLink
             type={instantiatorType}
-            value={instantiateInfo.instantiator}
+            value={contractDetail.instantiator ?? "N/A"}
             showCopyOnHover
             fixedHeight
           />
@@ -206,14 +228,16 @@ export const InstantiateInfo = ({
             initTxHash={initTxHash}
             initProposalId={initProposalId}
             initProposalTitle={initProposalTitle}
-            createdHeight={instantiateInfo.createdHeight}
+            createdHeight={createdHeight}
           />
         </Flex>
       </Flex>
       <Flex direction={{ base: "row", md: "column" }} gap={{ base: 1, md: 6 }}>
-        {instantiateInfo.ibcPortId && (
+        {rawContractResponse?.contract_info.ibc_port_id && (
           <LabelText label="IBC Port ID">
-            <PortIdRender portId={instantiateInfo.ibcPortId} />
+            <PortIdRender
+              portId={rawContractResponse.contract_info.ibc_port_id}
+            />
           </LabelText>
         )}
       </Flex>